requirements.txt
asgiref==3.8.1
aspose-pdf==24.2.0    
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.11
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
Levenshtein==0.25.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pdf2image==1.17.0
pdfminer==20191125
pillow==10.2.0
pycryptodome==3.20.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12
threadpoolctl==3.3.0
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.17.0